session: migrate test-infra to testify for clustered_index_test (#28901)
